Moved Action class definition to separate module
