+++ pathfinding/pathfind.js
@@ -1,4 +1,5 @@
 import { dragElement, ResizeHandler } from "../js/draggableCard.js";
+import { Action } from "../js/Action.js";
 
 const speeds = [1, 2, 4]
 let selectedFunction = (new URLSearchParams(window.location.search)).get("func") || AStar
@@ -440,27 +441,6 @@ document.querySelector("#PlayPause").onclick = function() {
         playing = true
         currentAnim.play()
         this.firstChild.setAttribute("src", "../Assets/pause-fill.svg")
-    }
-}
-
-class Action {
-    constructor(targets, line) {
-        this.targets = targets
-        this.line = `#pseudo${line}`
-    }
-
-    get duration() {
-        return 1000
-    }
-
-    AnimatePseudocode() {
-        anime({
-            targets: this.line,
-            backgroundColor: [{value: "#000000", duration: this.duration-1},
-                {value: anime.get(document.querySelector(`${this.line}`), "backgroundColor"), duration: 1}],
-            color: [{value: "#FFFFFF", duration: this.duration-1},
-                {value: anime.get(document.querySelector(`${this.line}`), "color"), duration: 1}]
-        })
     }
 }
